Provide memory to `CPU` struct as a mutable reference to `u8` slice
This is to be able to provide a way to test instruction that require
data to be in memory somewhere, without using private methods on the
`CPU` struct in tests to perform the memory-writingm and also without
needing to use other instructions to get data into a memory address.

Instead, the memory array can be modifed as needed for the test upfront,
and then a mutable reference to the underlying slice is passed to the
`CPU` struct's `new()` method.