Fix incorrect stack memory address layout
Previously, the least significant byte of stack memory addresses were
hardcoded as `0x01`, and varying the stack pointer caused the most
significant byte of addresses in the stack to change.

This is exactly the opposite of what should be done, where the most
significant byte of memory addresses should be fixed to `0x01`, and
varying the stack pointer should affect the least significant byte of
stack memory addresses, rather than affecting the most significant byte.